@@ -1,31 +1,38 @@
 package com.moengage.flutter
 
+import android.content.ContentValues.TAG
 import com.moengage.core.internal.logger.Logger
 import com.moengage.plugin.base.EventEmitter
 import com.moengage.plugin.base.inAppCampaignToJson
-import com.moengage.plugin.base.model.Event
-import com.moengage.plugin.base.model.EventType
-import com.moengage.plugin.base.model.InAppEvent
-import com.moengage.plugin.base.model.PushEvent
+import com.moengage.plugin.base.model.*
 import com.moengage.plugin.base.pushPayloadToJson
+import com.moengage.plugin.base.pushTokenToJson
 import org.json.JSONObject
 import java.util.*
+
 
 /**
  * @author Arshiya Khanum
  * Date: 2020/10/21
  */
-class EventEmitterImpl : EventEmitter {
+class EventEmitterImpl(private val onEvent: (methodName: String, payload: String) -> Unit) :
+    EventEmitter {
 
     private val tag: String = "${MODULE_TAG}EventEmitterImpl"
 
     override fun emit(event: Event) {
         try {
             Logger.v("$tag emit() : $event")
-            if (event is InAppEvent) {
-                emitInAppEvent(event)
-            } else if (event is PushEvent) {
-                emitPushEvent(event)
+            when (event) {
+                is InAppEvent -> {
+                    emitInAppEvent(event)
+                }
+                is PushEvent -> {
+                    emitPushEvent(event)
+                }
+                is TokenEvent -> {
+                    emitPushTokenEvent(event)
+                }
             }
         } catch (e: Exception) {
             Logger.e("$tag emit() : Exception: ", e)
@@ -58,17 +65,29 @@ class EventEmitterImpl : EventEmitter {
         }
     }
 
+    private fun emitPushTokenEvent(tokenEvent: TokenEvent) {
+        try {
+            Logger.v("$TAG emitPushTokenEvent() : $tokenEvent")
+            val eventType =
+                eventMap[tokenEvent.eventType] ?: return
+            val payload = pushTokenToJson(tokenEvent.pushToken)
+            emit(eventType, payload)
+        } catch (e: Exception) {
+            Logger.e("$TAG emitPushTokenEvent() : ", e)
+        }
+    }
+
     private fun emit(methodName: String, payload: JSONObject) {
         try {
             Logger.v("$tag emit() : methodName: $methodName")
-            MoEngageFlutterPlugin.sendCallback(methodName, payload.toString())
+            onEvent(methodName, payload.toString())
         } catch (e: Exception) {
             Logger.e("$tag emit() : ", e)
         }
     }
 
     companion object {
-        
+
         private val eventMap = EnumMap<EventType, String>(EventType::class.java)
 
         init {
@@ -78,6 +97,7 @@ class EventEmitterImpl : EventEmitter {
             eventMap[EventType.INAPP_CLOSED] = "onInAppDismiss"
             eventMap[EventType.INAPP_CUSTOM_ACTION] = "onInAppCustomAction"
             eventMap[EventType.INAPP_SELF_HANDLED_AVAILABLE] = "onInAppSelfHandle"
+            eventMap[EventType.PUSH_TOKEN_GENERATED] = "onPushTokenGenerated"
         }
     }
 }
