@@ -14,6 +14,11 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 
 class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
 
+    private val tag = "${MODULE_TAG}MoEngageFlutterPlugin"
+    private lateinit var channel: MethodChannel
+    private lateinit var context: Context
+    private val pluginHelper = PluginHelper()
+
     override fun onAttachedToEngine(binding: FlutterPluginBinding) {
         Logger.v("$tag onAttachedToEngine() : Registering MoEngageFlutterPlugin")
         context = binding.applicationContext
@@ -21,8 +26,37 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
     }
 
     override fun onDetachedFromEngine(binding: FlutterPluginBinding) {
-        Logger.v("$tag onDetachedFromEngine() : Registering MoEngageFlutterPlugin")
-        pluginHelper.onFrameworkDetached()
+        try {
+            Logger.v("$tag onDetachedFromEngine() : Registering MoEngageFlutterPlugin")
+            pluginHelper.onFrameworkDetached()
+            channel.setMethodCallHandler(null)
+        } catch (e: Exception) {
+            Logger.e("$tag onDetachedFromEngine() ", e)
+        }
+    }
+
+    private fun initPlugin(binaryMessenger: BinaryMessenger) {
+        try {
+            channel = MethodChannel(binaryMessenger, FLUTTER_PLUGIN_CHANNEL_NAME)
+            channel.setMethodCallHandler(this)
+            pluginHelper.setEventCallback(EventEmitterImpl(::sendCallback))
+        } catch (ex: Exception) {
+            Logger.e("$tag initPlugin() : exception : ", ex)
+        }
+    }
+
+    private fun sendCallback(methodName: String, message: String) {
+        try {
+            Handler(Looper.getMainLooper()).post {
+                try {
+                    channel.invokeMethod(methodName, message)
+                } catch (e: Exception) {
+                    Logger.e("$tag sendCallback() ", e)
+                }
+            }
+        } catch (ex: Exception) {
+            Logger.e("$tag sendCallback() : exception: ", ex)
+        }
     }
 
     @Suppress("SENSELESS_COMPARISON")
@@ -58,6 +92,7 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
                 METHOD_NAME_PUSH_TOKEN -> passPushToken(call)
                 METHOD_NAME_OPT_OUT_TRACKING -> optOutTracking(call)
                 METHOD_NAME_SELF_HANDLED_CALLBACK -> selfHandledCallback(call)
+                METHOD_NAME_UPDATE_SDK_STATE -> updateSdkState(call)
                 else -> Logger.e("$tag onMethodCall() : No mapping for this method.")
             }
         } catch (e: Exception) {
@@ -203,42 +238,25 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
         }
     }
 
-    private fun selfHandledCallback(methodCall: MethodCall){
-        try{
+    private fun selfHandledCallback(methodCall: MethodCall) {
+        try {
             if (methodCall.arguments == null) return
             val payload = methodCall.arguments.toString()
             Logger.v("$tag selfHandledCallback() : Arguments: $payload")
             pluginHelper.selfHandledCallback(context, payload)
-        }catch(e: Exception){
+        } catch (e: Exception) {
             Logger.e("$tag selfHandledCallback() : ", e)
         }
     }
 
-
-    companion object {
-
-        private const val tag = "${MODULE_TAG}MoEngageFlutterPlugin"
-        private lateinit var context: Context
-        private var channel: MethodChannel? = null
-        private val pluginHelper = PluginHelper()
-
-        private fun initPlugin(binaryMessenger: BinaryMessenger) {
-            try {
-                channel = MethodChannel(binaryMessenger, FLUTTER_PLUGIN_CHANNEL_NAME)
-                val plugin = MoEngageFlutterPlugin()
-                channel?.setMethodCallHandler(plugin)
-                pluginHelper.setEventCallback(EventEmitterImpl())
-            } catch (ex: Exception) {
-                Logger.e("$tag initPlugin() : exception : ", ex)
-            }
-        }
-
-        fun sendCallback(methodName: String, message: String) {
-            try {
-                Handler(Looper.getMainLooper()).post { channel?.invokeMethod(methodName, message) }
-            } catch (ex: Exception) {
-                Logger.e("$tag sendCallback() : exception: ", ex)
-            }
+    private fun updateSdkState(methodCall: MethodCall) {
+        try {
+            if (methodCall.arguments == null) return
+            val payload = methodCall.arguments.toString()
+            Logger.v("$tag selfHandledCallback() : Arguments: $payload")
+            pluginHelper.storeFeatureStatus(context, payload);
+        } catch (e: Exception) {
+            Logger.e("$tag selfHandledCallback() : ", e)
         }
     }
 }
