@@ -10,17 +10,22 @@ import MoEngage
 import UserNotifications
 import MoEPluginBase
 
-public class MOFlutterInitializer : NSObject {
+@objc public class MOFlutterInitializer : NSObject {
     
-    static public let sharedInstance = MOFlutterInitializer()
+    @objc static public let sharedInstance = MOFlutterInitializer()
     private override init() {super.init()}
     
-    public func initializeWithAppID(_ appID: String, withLaunchOptions options: [UIApplication.LaunchOptionsKey: Any]?) {
-        // Track Integration type
-        MoEPluginBridge.sharedInstance()?.trackPluginVersion(MOFlutterPluginInfo.kVersion, for: Flutter)
+    @objc public func initializeWithSDKConfig(_ config: MOSDKConfig, andLaunchOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
+        self.initializeWithSDKConfig(config, withSDKState: MoEngageCore.sharedInstance().isSDKEnabled(), andLaunchOptions: launchOptions)
+    }
+    
+    @objc public func initializeWithSDKConfig(_ config: MOSDKConfig, withSDKState state:Bool, andLaunchOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
+        // Track Integration type and version
+        config.pluginIntegrationType = FLUTTER
+        config.pluginIntegrationVersion = MOFlutterPluginInfo.kVersion
         
         // Initialize SDK
-        MoEPluginInitializer.sharedInstance().intializeSDK(withAppID: appID, andLaunchOptions: options ?? [:])
+        MoEPluginInitializer.sharedInstance().intializeSDK(with: config, withSDKState: state, andLaunchOptions: launchOptions ?? [:])
     }
     
 }
