@@ -1,5 +1,5 @@
 // Generated file, do not edit
 import Foundation
 struct MOFlutterPluginInfo{
- static let kVersion = "2.0.3" 
+ static let kVersion = "3.0.1" 
 }
