@@ -19,6 +19,7 @@ const String methodResetAppContext = "resetCurrentContext";
 const String methodOptOutTracking = "optOutTracking";
 const String methodEnableSDKLogs = "enableSDKLogs";
 const String methodLogout = "logout";
+const String methodUpdateSdkState = "updateSdkState";
 
 // Callback Names
 const String callbackOnPushClick = "onPushClick";
@@ -27,6 +28,7 @@ const String callbackOnInAppClicked = "onInAppClick";
 const String callbackOnInAppDismissed = "onInAppDismiss";
 const String callbackOnInAppCustomAction = "onInAppCustomAction";
 const String callbackOnInAppSelfHandled = "onInAppSelfHandle";
+const String callbackPushTokenGenerated = "onPushTokenGenerated";
 
 // Gender Value Constants
 const String genderMale = "male";
@@ -57,13 +59,14 @@ const String keyAttrLatitudeName = "latitude";
 const String keyAttrLongitudeName = "longitude";
 const String keyPushToken = "token";
 const String keyPushPayload = "pushPayload";
-const String keyPushService = "service";
+const String keyService = "service";
 const String keyAttributeType = "type";
 const String keyAlias = "alias";
 const String keyLocationAttribute = "locationAttribute";
 const String keyState = "state";
 const String keyAppStatus = "appStatus";
 const String keyContexts = "contexts";
+const String keyPushService = "pushService";
 
 const String keyPayload = "payload";
 const String keyKvPair = "kvPair";
@@ -98,13 +101,11 @@ const String gdprOptOutTypeData = "data";
 const String gdprOptOutTypePush = "push";
 const String gdprOptOutTypeInApp = "inapp";
 
-// Push Service Types
-const String pushServiceFCM = "FCM";
-const String pushServicePushKit = "PUSH_KIT";
-const String pushServiceMiPush = "MI_PUSH";
-
 // SelfHandled Callback Action Types
 const String selfHandledActionShown = "impression";
 const String selfHandledActionPrimaryClicked = "primary_clicked";
 const String selfHandledActionClick = "click";
 const String selfHandledActionDismissed = "dismissed";
+
+// SDK Status update
+const String keyIsSdkEnabled = "isSdkEnabled";
